Refactor conversations API with modular router
- Moved email conversations logic to a dedicated `server-api.js` file for better modularity.
- Introduced `createConversationsRouter` for managing conversation endpoints.
- Simplified `index.js` by replacing inlined conversation logic with reusable router.
